upgrade web UI

+++ web/app/static/js/themeToggle.js
@@ -1,19 +1,23 @@
-document.getElementById('dark-mode-toggle').addEventListener('click', function() {
-    const isDarkMode = document.body.classList.toggle('dark-mode');
-    
-    // Change the icon based on the current mode
-    const icon = isDarkMode ? 'sun-icon.svg' : 'moon-icon.svg';
-    this.querySelector('img').src = `{{ url_for('static', filename='images/${icon}') }}`;
-    
-    // Save mode in localStorage
-    localStorage.setItem('theme', isDarkMode ? 'dark' : 'light');
-});
+document.addEventListener('DOMContentLoaded', function() {
+    const darkModeToggle = document.getElementById('dark-mode-toggle');
 
-// Apply saved theme on page load
-window.addEventListener('DOMContentLoaded', (event) => {
-    const savedTheme = localStorage.getItem('theme');
-    if (savedTheme === 'dark') {
-        document.body.classList.add('dark-mode');
-        document.getElementById('dark-mode-toggle').querySelector('img').src = '/static/images/sun-icon.svg';
+    if (darkModeToggle) {  // Ensure the toggle button exists
+        darkModeToggle.addEventListener('click', function() {
+            const isDarkMode = document.body.classList.toggle('dark-mode');
+            
+            // Update the icon based on the current mode
+            const icon = isDarkMode ? 'sun-icon.svg' : 'moon-icon.svg';
+            this.querySelector('img').src = `/static/images/${icon}`;
+            
+            // Save the mode in localStorage
+            localStorage.setItem('theme', isDarkMode ? 'dark' : 'light');
+        });
+
+        // Apply the saved theme when the page loads
+        const savedTheme = localStorage.getItem('theme');
+        if (savedTheme === 'dark') {
+            document.body.classList.add('dark-mode');
+            darkModeToggle.querySelector('img').src = '/static/images/sun-icon.svg';
+        }
     }
 });
